requirement.txt
streamlit
pandas
langchain_google_genai
beautifulsoup4
langchain
langchain-community
langchain-google-genai
langchain-core
python-dotenv
PyPDF


